requirements_test.txt
pytest-homeassistant-custom-component
pytest-asyncio
syrupy
mypy
ruff
